Update JuMP and linopy benchmark scripts
This commit slightly tweaks the JuMP and linopy benchmark scripts.

The JuMP models are rewritten to be more similar to the style of linopy, even
though it is slightly slower.

The linopy objective is tweaked to result in a single multiplication instead of
n^2.

In future, the memory usage of the JuMP model could be reduced by using
m=direct_model(solver()). This is  not the default because it does not work for
all solvers (like Cbc).